requirements.txt
streamlit
pandas
pydantic
langchain-openai
langchain-core
python-dotenv